feat: add Mermaid exporter and TLD metadata support
